HAJAE-9 rename to kebab case
