+++ src/app/(pages)/about/_components/side-nav/side-nav.module.scss
