+++ src/app/_components/snake-game/snake-game.module.scss
