+++ src/enums/enum-base.ts
