+++ src/lib/features/about-explorer-slice.ts
