+++ src/ui/icon-box/icon-box.tsx
