+++ src/ui/technologies-row/technologies-row.module.scss
