+++ src/ui/technology-logo/technology-logo.module.scss
